Tratamento de Exceções

+++ src/main/java/com/example/meu_primeiro_springboot/service/ProdutoService.java
@@ -1,5 +1,6 @@
 package com.example.meu_primeiro_springboot.service;
 
+import com.example.meu_primeiro_springboot.exceptions.RecursoNaoEncontradoException;
 import com.example.meu_primeiro_springboot.model.Produto;
 import com.example.meu_primeiro_springboot.repository.ProdutoRepository;
 import org.springframework.stereotype.Service;
@@ -20,8 +21,9 @@ public class ProdutoService {
     return produtoRepository.findAll();
   }
 
-  public Optional<Produto> buscarProduto(Long id){
-    return produtoRepository.findById(id);
+  public Produto buscarProduto(Long id){
+    return produtoRepository.findById(id)
+        .orElseThrow(() -> new RecursoNaoEncontradoException("Produto com ID "+id+" não encontrado"));
   }
 
   public Produto salvarProduto(Produto produto){
@@ -29,6 +31,10 @@ public class ProdutoService {
   }
 
   public void deletarProduto(Long id){
+
+      if (!produtoRepository.existsById(id)){
+        throw new RecursoNaoEncontradoException("Produto com ID"+id+" não encontrado");
+      }
      produtoRepository.deleteById(id);
   }
 }
